mark: move marking good/bad/active into new file
This commit is preliminary with regard to the imminent new D-Bus method 'mark'
insofar that the functionality is prepared for it and is given its own source
file.
