+++ include/mark.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <glib.h>
+
+gboolean mark_run(const gchar *state,
+		  const gchar *slot_identifier,
+		  gchar **slot_name,
+		  gchar **message);
